refactor(ui): split ContextMenu into separate component files
